core/z80: Fixed IX+d and IY+d addressing.

+++ src/core/zilog_z80_i_ix.cpp
@@ -824,7 +824,7 @@ int Z80::i_ADC_A_mIXd(Z80& self, uint8_t op, addr_t arg)
      * N is reset.
      * C is set if carry from bit 7; otherwise, it is reset.
      */
-    addr_t addr = self._regs.IX + arg;
+    addr_t addr = self._regs.IX + static_cast<int8_t>(arg);
     uint8_t value = self.read(addr);
     self._regs.memptr = addr;
     return self.add_A(value, self.test_C());
@@ -845,7 +845,7 @@ int Z80::i_SUB_A_mIXd(Z80& self, uint8_t op, addr_t arg)
     /*
      * SUB (IX+d)
      */
-    addr_t addr = self._regs.IX + arg;
+    addr_t addr = self._regs.IX + static_cast<int8_t>(arg);
     int value = self.read(addr);
     self._regs.memptr = addr;
     return self.sub_A(value, 0);
@@ -866,7 +866,7 @@ int Z80::i_SBC_A_mIXd(Z80& self, uint8_t op, addr_t arg)
     /*
      * SBC (IX+d)
      */
-    addr_t addr = self._regs.IX + arg;
+    addr_t addr = self._regs.IX + static_cast<int8_t>(arg);
     uint8_t value = self.read(addr);
     self._regs.memptr = addr;
     return self.sub_A(value, self.test_C());
@@ -886,7 +886,7 @@ int Z80::i_AND_A_mIXd(Z80& self, uint8_t op, addr_t arg)
     /*
      * AND (IX+d)
      */
-    addr_t addr = self._regs.IX + arg;
+    addr_t addr = self._regs.IX + static_cast<int8_t>(arg);
     uint8_t value = self.read(addr);
     self._regs.memptr = addr;
     return self.and_A(value);
@@ -906,7 +906,7 @@ int Z80::i_XOR_A_mIXd(Z80& self, uint8_t op, addr_t arg)
     /*
      * XOR (IX+d)
      */
-    addr_t addr = self._regs.IX + arg;
+    addr_t addr = self._regs.IX + static_cast<int8_t>(arg);
     uint8_t value = self.read(addr);
     self._regs.memptr = addr;
     return self.xor_A(value);
@@ -926,7 +926,7 @@ int Z80::i_OR_A_mIXd(Z80& self, uint8_t op, addr_t arg)
     /*
      * OR (IX+d)
      */
-    addr_t addr = self._regs.IX + arg;
+    addr_t addr = self._regs.IX + static_cast<int8_t>(arg);
     uint8_t value = self.read(addr);
     self._regs.memptr = addr;
     return self.or_A(value);
@@ -946,7 +946,7 @@ int Z80::i_CP_A_mIXd(Z80& self, uint8_t op, addr_t arg)
     /*
      * CP (IX+d)
      */
-    addr_t addr = self._regs.IX + arg;
+    addr_t addr = self._regs.IX + static_cast<int8_t>(arg);
     uint8_t value = self.read(addr);
     self._regs.memptr = addr;
     return self.cp_A(value);
